+++ dukes-tomee/src/main/java/eu/agilejava/dukes/package-info.java
@@ -21,21 +21,4 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package eu.agilejava.snoop;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.core.Response;
-
-/**
- *
- * @author Ivar Grimstad (ivar.grimstad@gmail.com)
- */
-@Path("hello")
-public class HelloResource {
-   
-   @GET
-   public Response greet() {
-      return Response.ok("Hello World! ...from Tomee Micro").build();
-   }
-}
+package eu.agilejava.dukes;
